EZP-30139: Update db data type

+++ data/update/mysql/dbupdate-7.4.0-to-7.5.0.sql
@@ -35,4 +35,4 @@ ALTER TABLE `eznotification` MODIFY COLUMN `data` TEXT;
 -- EZP-30139: As an editor I want to hide and reveal a content item
 --
 
-ALTER TABLE `ezcontentobject` ADD COLUMN `is_hidden` INT(11) NOT NULL DEFAULT '0';
+ALTER TABLE `ezcontentobject` ADD COLUMN `is_hidden` tinyint(1) NOT NULL DEFAULT '0';
